Initial import of project.
Trying to write a wav type file but AudioFileInitializeWithCallbacks
always fails to return a valid AudioFileID. It appears CoreAudio tries
to write the file then read it back to check itself. This is difficult
to do as we only have access to the OutputStream. I have tried some
horrible hacks by adding all the data written to a MemoryBlock but
still can't get a valid AudioFileID.